+++ code/cell.modules/pkg.sys/ui.harness/src/api/action/ActionBuilder.TEST.tsx
@@ -14,7 +14,7 @@ function create() {
   return { model, builder };
 }
 
-describe.only('ActionBuilder', () => {
+describe('ActionBuilder', () => {
   describe('ActionBuilder.model()', () => {
     it('model', () => {
       const model = ActionBuilder.model();
