+++ code/cell.modules/pkg.sys/ui.harness/src/api/index.ts
@@ -1,1 +1,1 @@
-export * from './ActionBuilder';
+export * from './action';
