+++ code/cell.modules/pkg.sys/ui.harness/src/test/Dev/Dev.actions.ts
@@ -31,7 +31,7 @@ export const actions = Actions<Ctx>()
 
   .button('change text', (ctx) => change(ctx.model))
   .button((config) => config.label('hello'))
-  .hr()
+  .hr(1, 0.2)
   .button('console.log', (ctx) => console.log('hello', ctx))
 
   .hr()
